dal v2

+++ src/eu/unareil/dal/jdbc/AuteursJDBC.java
@@ -98,7 +98,7 @@ public class AuteursJDBC implements DAO<Auteur> {
     }
 
     @Override
-    public Auteur selectById(int id) throws DALException {
+    public Auteur selectById(long id) throws DALException {
         PreparedStatement pstmt = null;
         ResultSet rs;
         Auteur el = null;
